server/requirements.txt
python-dotenv
openai==0.28.0
Flask==2.2.3
flask_cors==4.0.0
dnspython
pymongo==4.5.0
gunicorn